requirements.txt
fire
pyyaml
firebase-admin